// apps/web/src/app/api-client/models/data-relation-entity.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DataRelationEntity {

  /**
   * Cardinalite de la relation.
   */
  cardinalite: 'UN_A_UN' | 'UN_A_PLUSIEURS' | 'PLUSIEURS_A_PLUSIEURS';

  /**
   * Identifiant de la relation.
   */
  id: string;

  /**
   * Libelle de la relation.
   */
  label?: string | null;

  /**
   * Identifiant de l'entite source.
   */
  sourceId: string;

  /**
   * Identifiant de l'entite cible.
   */
  targetId: string;
}